Working on Sword attacks

+++ Hack-and-Slash Game/Assets/Scripts/Enemies/EnemyData.cs
@@ -4,12 +4,12 @@ using System;
 [Serializable]
 public class EnemyStats
 {
-	public int health;
+	public float health;
 	public float damage;
 	public float speed;
 	public float timeToAttack;
 	
-	public EnemyStats(int health, float damage, float speed, float timeToAttack)
+	public EnemyStats(float health, float damage, float speed, float timeToAttack)
 	{
 		this.health = health;
 		this.damage = damage;
